MVP v1
This is a completely functioning prototype.  We have the code for the server, the client, and the RUST translation layer.  We are about to start implementing major changes across the code to bring in more features.  This is the first state of the MVP where everything works well enough to demonstrate this as a concept to others.